ai-hedge-fund/requirements.txt
# AI-Native Hedge Fund — Dependencies

# Core trading
backtrader==1.9.78.123
alpaca-trade-api==3.3.2
alpaca-py==0.26.0

# Options pricing & Greeks
QuantLib==1.35

# Anthropic Claude AI
anthropic==0.40.0

# Data & math
numpy==1.26.4
pandas==2.2.2
scipy==1.13.0

# Environment & utilities
python-dotenv==1.0.1
requests==2.32.3
websocket-client==1.8.0
schedule==1.2.2

# Logging & monitoring
rich==13.7.1
loguru==0.7.2

# Optional: Jupyter for analysis
jupyter==1.1.1
matplotlib==3.9.0
